backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
argon2-cffi==23.1.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-magic==0.4.27
pillow==10.1.0

# Environment & Config
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Async & Performance
aiofiles==23.2.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3